+++ src/main/java/io/github/mateuszuran/PTD/Manager/Trip/TripController.java
@@ -47,17 +47,26 @@ public class TripController {
     }
 
     @GetMapping("/trip/edit/{id}")
-    public String showEditTripForm(@PathVariable("id") Integer id, Model model) {
+    public String showEditTripForm(@PathVariable("id") Integer id, Model model, @AuthenticationPrincipal CustomUserDetails userDetails) {
+        Card cardId = cardService.findByUserId(userDetails.getUserId());
+        if(cardService.checkIfCardIsDone(userDetails.getUserId())) {
+            return "redirect:/home/card/" + cardId.getId() + "/?false";
+        } else {
         Trip trip = tripService.findTripById(id);
         model.addAttribute("trip", trip);
         return "trip_form";
+        }
     }
 
     @GetMapping("/trip/delete/{id}")
     public String deleteTrip(@PathVariable("id") Integer id, @AuthenticationPrincipal CustomUserDetails userDetails) {
         Card cardId = cardService.findByUserId(userDetails.getUserId());
-        tripService.deleteTrip(id);
-        counterService.toggleToFalse(cardId.getId());
-        return "redirect:/home/card/" + cardId.getId();
+        if(cardService.checkIfCardIsDone(userDetails.getUserId())) {
+            return "redirect:/home/card/" + cardId.getId() + "/?false";
+        } else {
+            tripService.deleteTrip(id);
+            counterService.toggleToFalse(cardId.getId());
+            return "redirect:/home/card/" + cardId.getId();
+        }
     }
 }
